gateway/pyproject.toml
[tool.poetry]
name = "gateway"
version = "0.1.0"
description = ""
authors = ["Christopher Zumberge <chriszumberge@gmail.com>"]
readme = "README.md"
packages = [
  { include = "gateway", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.1.0"
pydantic = "^2.11.3"
httpx = "^0.28.1"
python-dotenv = "^1.1.0"
boto3 = "^1.34.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[tool.poetry.scripts]
gateway = "gateway.app:main"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
